Moved chart models
