+++ Willowcat.CharacterGenerator.Core.Tests/Extension/OptionCollectionExtension.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using Willowcat.CharacterGenerator.Core.Models;
 using Willowcat.CharacterGenerator.Model;
 
 namespace Willowcat.CharacterGenerator.Core.Tests.Extension
