+++ Willowcat.CharacterGenerator.Core/Randomizer/RandomNameChart.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using Willowcat.CharacterGenerator.Core.Randomizer;
+using Willowcat.CharacterGenerator.Model;
 
 namespace Willowcat.CharacterGenerator.Core.Models
 {
