+++ Willowcat.CharacterGenerator.Core/TextRepository/ChartParsingException.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Willowcat.CharacterGenerator.Core.Models;
+using Willowcat.CharacterGenerator.Model;
 
 namespace Willowcat.CharacterGenerator.Core.TextRepository
 {
